multiply args

+++ 0x0A-argc_argv/3-mul.c
@@ -9,7 +9,7 @@
 
 int main(int argc, char *argv[])
 {
-	int product = 1;
+	long int product = 1;
 	int i, n;
 
 	if (argc < 2 || argc > 3)
@@ -23,7 +23,7 @@ int main(int argc, char *argv[])
 			n = strtol(argv[i], NULL, 10);
 			product *= n;
 		}
-		printf("%d\n", product);
+		printf("%ld\n", product);
 	}
 	return (0);
 }
